backend/requirements.txt
fastapi==0.111.1
httpx==0.27.0
PyJWT==2.8.0
mysqlclient==2.2.4
pydantic==2.8.2
pinecone-client==5.0.0
sentence-transformers==2.2.2
sqlalchemy==1.4.27
uvicorn==0.30.3
psycopg2==2.9.9
python-dotenv==0.19.1
huggingface_hub==0.24.0